examples/images_preprocessing/requirements.txt
zenml
pillow
click